Rename stuff and change process to be static instead of only ID
