+++ AzzyBot/Modules/MusicStreaming/MusicStreamingModule.cs
@@ -28,12 +28,12 @@ internal sealed class MusicStreamingModule : BaseModule
 
     internal override async void StartProcesses()
     {
-        if (!await MusicStreamingLavalink.CheckIfJavaIsInstalledAsync())
+        if (!await MusicStreamingLavalinkHandler.CheckIfJavaIsInstalledAsync())
             throw new InvalidOperationException("You have to install Java/OpenJDK Runtime 17 or 21 first!");
 
-        MusicStreamingLavalink.StartLavalink();
+        MusicStreamingLavalinkHandler.StartLavalink();
     }
 
-    internal override async void StopProcesses() => await MusicStreamingLavalink.StopLavalinkAsync();
+    internal override async void StopProcesses() => await MusicStreamingLavalinkHandler.StopLavalinkAsync();
     internal override void Activate() => ModuleStates.ActivateMusicStreaming();
 }
